--- a/stacktrace.go
+++ b/stacktrace.go
@@ -31,12 +31,6 @@ import (
 const _zapPackage = "go.uber.org/zap"
 
 var (
-	_stacktracePool = sync.Pool{
-		New: func() interface{} {
-			return newProgramCounters(64)
-		},
-	}
-
 	// We add "." and "/" suffixes to the package name to ensure we only match
 	// the exact package and not any package with the same prefix.
 	_zapStacktracePrefixes       = addPrefix(_zapPackage, ".", "/")
@@ -44,27 +38,19 @@ var (
 )
 
 func takeStacktrace() string {
+	pcs := newProgramCounters()
+	defer pcs.Release()
+
+	return makeStacktrace(pcs.Callers(1))
+}
+
+func makeStacktrace(pcs []uintptr) string {
 	buffer := bufferpool.Get()
 	defer buffer.Free()
-	programCounters := _stacktracePool.Get().(*programCounters)
-	defer _stacktracePool.Put(programCounters)
-
-	var numFrames int
-	for {
-		// Skip the call to runtime.Counters and takeStacktrace so that the
-		// program counters start at the caller of takeStacktrace.
-		numFrames = runtime.Callers(2, programCounters.pcs)
-		if numFrames < len(programCounters.pcs) {
-			break
-		}
-		// Don't put the too-short counter slice back into the pool; this lets
-		// the pool adjust if we consistently take deep stacktraces.
-		programCounters = newProgramCounters(len(programCounters.pcs) * 2)
-	}
 
 	i := 0
 	skipZapFrames := true // skip all consecutive zap frames at the beginning.
-	frames := runtime.CallersFrames(programCounters.pcs[:numFrames])
+	frames := runtime.CallersFrames(pcs)
 
 	// Note: On the last iteration, frames.Next() returns false, with a valid
 	// frame, but we ignore this frame. The last frame is a a runtime frame which
@@ -113,8 +99,34 @@ type programCounters struct {
 	pcs []uintptr
 }
 
-func newProgramCounters(size int) *programCounters {
-	return &programCounters{make([]uintptr, size)}
+var _stacktracePool = sync.Pool{
+	New: func() interface{} {
+		return &programCounters{make([]uintptr, 64)}
+	},
+}
+
+func newProgramCounters() *programCounters {
+	return _stacktracePool.Get().(*programCounters)
+}
+
+func (pcs *programCounters) Release() {
+	_stacktracePool.Put(pcs)
+}
+
+func (pcs *programCounters) Callers(skip int) []uintptr {
+	var numFrames int
+	for {
+		// Skip the call to runtime.Callers and programCounter.Callers
+		// so that the program counters start at our caller.
+		numFrames = runtime.Callers(skip+2, pcs.pcs)
+		if numFrames < len(pcs.pcs) {
+			return pcs.pcs[:numFrames]
+		}
+
+		// Don't put the too-short counter slice back into the pool; this lets
+		// the pool adjust if we consistently take deep stacktraces.
+		pcs.pcs = make([]uintptr, len(pcs.pcs)*2)
+	}
 }
 
 func addPrefix(prefix string, ss ...string) []string {